refactor(cli): extract artifact serialization
